Minor comments
